# swagger.yaml
openapi: 3.0.3
info:
  title: Chatbot Server API
  description: Customer support chatbot system with real-time messaging and order/product management
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /api/tickets:
    get:
      tags:
        - Tickets
      summary: Get support tickets
      description: Retrieve support tickets with optional filtering
      parameters:
        - name: status
          in: query
          description: Filter by ticket status
          schema:
            type: string
            enum: [open, assigned, resolved, closed]
        - name: agentId
          in: query
          description: Filter by assigned agent ID
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of tickets to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Tickets retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/TicketSummary'
                  total:
                    type: integer
                    description: Number of tickets returned
  /api/tickets/{ticketId}:
    get:
      tags:
        - Tickets
      summary: Get ticket details
      description: Retrieve detailed information about a specific support ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          description: Ticket ID
          schema:
            type: string
          example: "TICKET-1628097234567-1"
      responses:
        '200':
          description: Ticket details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket:
                    $ref: '#/components/schemas/Ticket'
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Agent management endpoints
...
components:
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
          description: Unique ticket identifier
        customerId:
          type: string
          description: Customer identifier
        customerSocketId:
          type: string
          description: Customer socket ID
        status:
          type: string
          enum: [open, assigned, resolved, closed]
          description: Ticket status
        priority:
          type: string
          enum: [low, normal, high, urgent]
          description: Ticket priority
        assignedAgent:
          type: string
          nullable: true
          description: ID of assigned agent
        createdAt:
          type: string
          format: date-time
          description: Ticket creation time
        updatedAt:
          type: string
          format: date-time
          description: Last update time
        closedAt:
          type: string
          format: date-time
          nullable: true
          description: Ticket closure time
        closedBy:
          type: string
          nullable: true
          description: Who closed the ticket
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        customerInfo:
          type: object
          properties:
            connectedAt:
              type: string
              format: date-time
            lastSeen:
              type: string
              format: date-time
    TicketSummary:
      type: object
      properties:
        id:
          type: string
          description: Ticket ID
        customerId:
          type: string
          description: Customer ID
        status:
          type: string
          enum: [open, assigned, resolved, closed]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: integer
          description: Total number of messages
        unreadCount:
          type: integer
          description: Number of unread messages
        lastMessage:
          type: object
          nullable: true
          properties:
            content:
              type: string
              description: Last message content (truncated)
            timestamp:
              type: string
              format: date-time
            sender:
              type: string
              enum: [customer, agent, system]
    Message:
      type: object
      properties:
        id:
          type: string
          description: Message ID
        content:
          type: string
          description: Message content
        sender:
          type: string
          enum: [customer, agent, system]
          description: Who sent the message
        senderId:
          type: string
          nullable: true
          description: ID of the sender
        timestamp:
          type: string
          format: date-time
          description: Message timestamp
        read:
          type: boolean
          description: Whether message has been read
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error

  # WebSocket Events Documentation (for reference)
  # Note: OpenAPI 3.0 doesn't support WebSocket documentation, but here are the events:
  # 
  # Customer Namespace (/customer):
  # Events Received:
  # - chat_response: { message, timestamp, messageId, sender, messageType }
  # - typing: { isTyping }
  # 
  # Events Sent:
  # - chat_message: { message }
  # - typing: { isTyping }
  # 
  # Agent Namespace (/agent):
  # Events Received:
  # - auth_success: { agentId, agentName, department, tickets, stats }
  # - auth_error: { message }
  # - new_ticket: { ticket, message }
  # - customer_message: { ticketId, message, customerInfo }
  # - customer_typing: { ticketId, isTyping }
  # - customer_disconnected: { ticketId, disconnectedAt }
  # - message_sent: { ticketId, messageId, timestamp }
  # - ticket_updated: { ticketId, status, updatedAt }
  # - ticket_details: { ticket }
  # - agent_status_change: { agentId, agentName, status }
  # 
  # Events Sent:
  # - authenticate: { agentId, agentName, department }
  # - send_message: { ticketId, message }
  # - typing: { ticketId, isTyping }
  # - update_ticket_status: { ticketId, status }
  # - update_status: { status }
  # - get_ticket_details: { ticketId }